Revert "chess fixes?"
This reverts commit a195612d19a6315a4ffded9466f203badda2260f.

+++ applications/chess/fast_chess.c
@@ -269,7 +269,7 @@ int toMinFen(char* fen, Position* position) {
             emptyCount++;
         } else {
             if(emptyCount != 0) {
-                snprintf(&fen[charCount++], 2, "%d", (int) emptyCount);
+                snprintf(&fen[charCount++], 2, "%d", emptyCount);
                 emptyCount = 0;
             }
             fen[charCount++] = bb2char(bb, &(position->board));
@@ -278,7 +278,7 @@ int toMinFen(char* fen, Position* position) {
         file++;
         if(file > 7) {
             if(emptyCount != 0) {
-                snprintf(&fen[charCount++], 2, "%d", (int) emptyCount);
+                snprintf(&fen[charCount++], 2, "%d", emptyCount);
                 emptyCount = 0;
             }
             file = 0;
